memories/requirements.txt
# geos
# proj
# gdal
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==3.2
django-allauth==0.50.0
django-filter==21.1
django-leaflet==0.28.3
djangorestframework==3.13.1
djangorestframework-gis==1.0
gensim==4.2.0
gnureadline==8.1.2
gunicorn==20.0.4
idna==3.3
numpy==1.22.4
oauthlib==3.2.0
postgis==1.0.4
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.0
requests-oauthlib==1.3.1
scipy==1.8.1
smart-open==6.0.0
sqlparse==0.4.2
urllib3==1.26.9
